Unify test_queryset_facets with test_facets
